move queue clearing code into environment.py
