+++ features/environment.py
@@ -1,9 +1,22 @@
+import json
 import logging
+import time
 import uuid
 from datetime import datetime
 
+import requests
+
 from config import Config
 from utilties.rabbit_context import RabbitContext
+
+
+def get_msg_count(queue_name):
+    uri = f'http://{Config.RABBITMQ_HOST}:{Config.RABBITMQ_MAN_PORT}/api/queues/%2f/{queue_name}'
+    response = requests.get(uri, auth=(Config.RABBITMQ_USER, Config.RABBITMQ_PASSWORD))
+    response.raise_for_status()
+    response_data = json.loads(response.content)
+
+    return response_data['messages']
 
 
 def before_all(_):
@@ -15,16 +28,37 @@ def before_all(_):
 def before_scenario(context, _):
     context.test_start_local_datetime = datetime.now()
     context.action_plan_id = str(uuid.uuid4())
-    _purge_queues()
+    _clear_down_all_queues()
 
 
-def _purge_queues():
-    with RabbitContext() as rabbit:
-        rabbit.channel.queue_purge(queue=Config.RABBITMQ_RH_OUTBOUND_CASE_QUEUE)
-        rabbit.channel.queue_purge(queue=Config.RABBITMQ_RH_OUTBOUND_UAC_QUEUE)
-        rabbit.channel.queue_purge(queue=Config.RABBITMQ_SAMPLE_INBOUND_QUEUE)
-        rabbit.channel.queue_purge(queue=Config.RABBITMQ_OUTBOUND_FIELD_QUEUE)
-        rabbit.channel.queue_purge(queue=Config.RABBITMQ_UNADDRESSED_REQUEST_QUEUE)
+def _clear_down_all_queues():
+    all_queues = _get_all_queues()
+
+    for q in all_queues:
+        # keep killing this Delayed queue, just to stop it redlivering anything in some mad race condition
+        _clear_down_queue(Config.RABBITMQ_DELAYED_REDELIVERY_QUEUE)
+        _clear_down_queue(q)
+
+
+def _get_all_queues():
+    uri = f'http://{Config.RABBITMQ_HOST}:{Config.RABBITMQ_MAN_PORT}/api/queues/%2f/'
+    response = requests.get(uri, auth=(Config.RABBITMQ_USER, Config.RABBITMQ_PASSWORD))
+    response.raise_for_status()
+    response_data = json.loads(response.content)
+
+    return [r['name'] for r in response_data]
+
+
+def _clear_down_queue(queue_name):
+    while True:
+        uri = f'http://{Config.RABBITMQ_HOST}:{Config.RABBITMQ_MAN_PORT}/api/queues/%2f/{queue_name}/contents'
+        response = requests.delete(uri, auth=(Config.RABBITMQ_USER, Config.RABBITMQ_PASSWORD))
+        response.raise_for_status()
+
+        if get_msg_count(queue_name) == 0:
+            return
+
+        time.sleep(1)
 
 
 def add_test_queue(binding_key, exchange_name, queue_name, exchange_type='topic'):
